Add basic neovim configuration with C++ support
The configuration is far from finished. It uses dein plugin manager and
some plugins to give a more IDE like C++ development experience.

This is the first commit of the new dotiles setup. It uses GNU stow to
manage all dotfiles from a centralized directory. This gives one a clear
overview of all dotfiles actively managed by the user and interworks
perfectly with version control systems.