Enable autocomplete on start-up
We currently use the plugin deoplete for autocompletion based on the
LanguageServer. On a long term it would be better to activate both the
LanguageClient plugin and deoplete only for those file types that are
supported (with the current configuration this would mean for C++
only). The plugins could then be lazy loaded when a supported file type
is opened.