Change linter to ale plugin
Previously, the linter of LanguageClient-neovim was used. We changed to
ale because it not only supports the LanguageServer protocol, but a lot
of different linters that can be used in parallel. Currently, only
cquery together with clang-tidy is activated for C++ files.

LanguageClient-neovim is still used for the other LanguageClient
functionalities such as "goto definition", where it has a way more
complete support for cquery in comparison to ale.