New nvim plugin handling & vim-unimpaired
I feel quite uncomfortable using dein.vim as plugin manager because
I do not have a feeling for its inner workings and the usage of the
interface is quite difficult to remember.

I therefore plan to switch to the package functionality added to
vim 8. The plugins then are added directly as subrepositories.

As a starting point I added vim-unimpaired as a new plugin using
the just described way of working.