Use bash completion for vim :Make command
The bash completion is provided by a vim plugin written by myself and
put to Github as a separate project. It is added to the dotfiles as a
submodule.

Further, the :Make command was changed to use the bash completion
provided by the plugin as completion. This provides a convenient way to
complete make targets in vim.