Solve non-working dynamic loading of vim-lsc
The language servers are only registered in after/plugin/lsc.vim which
is not sourced when using packadd dynamically on loading of source
files. We now register the language server manually upon loading of C++
source files. With this addition, the packadd approach to loading
plugins works with vim-lsc as well.