Extract file specific logic for C++ to vim ftplugin file
This is the recommended way to work with filetype specific
configurations and keeps the main configuration file smaller.

Plugins that are only needed for the cpp filetype will still be only
dynamically loaded when a C/C++ file is opened. This keeps startup times
as short as possible.

Somehow, the vim-lsc plugin does not work natively with this setup as it
makes use of the plugin/after functionality which was not intended to
be used by plugin authors to my understanding.

This leads to the language server not getting activated while loading a
C/C++ file. As a workaround the activation function for the language
server is called by hand in the fplugin file. The same applies for the
registration of the key mappings associated to the plugin. We add an
explicit call to the registration function as otherwise the key mappings
do not seem to be available for the first C/C++ file opended during a session.