By hand configuration: auto-formatting of cpp files in vim
Previously, this functionality was provided by the vim-autoformat
plugin.

However, this can be configured easily by hand. In the ftplugin files we
set the formatpgr for the buffers and use an autocommand to
automatically format the file before writing it to disk.

This commit therefore makes the vim-autoformat plugin obsolete and we
remove it.