Refined formatting configuration C++ files in vim
Previously, clang-format was just set as formatprg in vim. However,
somehow this did not work reliably for me if used in an auto-command
triggered on writing a buffer.

We now use clang-format via a buffer local formatexpr setting. This has
the advantage that we now also can only format selected ranges instead
of only the whole file.