Use vim-projectionist for alternate files
Previously, vim-altr was used to switch between source and header files
for C++ projects. For certain projects we also want to switch to certain
other related files like UT files or mocks.

This is usually project specific. To have this higher flexibility we
switch to the vim-projectionist plugin from vim-altr.