Change default window manager from herbstluftwm to dwm
Currently, both the statusbar script and the configuration of dwm are
not yet under version control.

For dwm we still have to find a good way to handle its special way of
configuration via a header file and recompiling upon configuration
changes.

For the statusbar script still some adaptations and a proper location on
the filesystem are needed.