LSP support in vim is handled by ALE plugin
Previously, the LSP support was handled by the separate plugin vim-lsc.
By now, most of the functionality for the language server protocol can
be handled by ALE directly.

A further advantage of this approach is that the language servers can be
directly reused for linting information.

For this to work with all the desired features an update of ALE to the
latest version was needed.