Neoterm windows in vim are opened in right split
Currently, the neoterm window is opened in the active window and deletes
the buffer in the active window. When running tests with vim-test this
is completely contrary to the desired behaviour as the active window
usually contains the testing code.

Instead, we now open a new split window on the right in which the tests
are run.