Replace python-lsp-server with jedi-language-server
The previously used python-lsp-server provides a lot of features via
plugins I do not need. In particular, it has plugins for linters I
prefer to call more directly via a dedicated plugin or from the
terminal.

Hence, I switch to jedi-language-server which provides a more
minimalistic set of features concentrating on the core functionality of
a language server.