Fix missing lsp config for null-ls only buffers
Previously the on_attach function was not executed upon registering the
sources for null-ls. As a consequence buffers which only used sources from
null-ls but no language server were not configured, i.e. mappings and
auto-formatting were missing. Explicitly, this was currently the case
for Lua files which were not auto-formatted on save.